fixed cyclic import error with getting solves and prblems
