+++ lbc/problems/problem.py
@@ -1,44 +1,7 @@
-from lbc.problems import Example0_lbc
-from lbc.problems.example1 import Example1
-from lbc.problems.example10 import Example10
-from lbc.problems.example11 import Example11
-from lbc.problems.example12 import Example12
-from lbc.problems.example2 import Example2
-from lbc.problems.example3 import Example3
-from lbc.problems.example4 import Example4
-from lbc.problems.example5 import Example5
-from lbc.problems.example6 import Example6
-from lbc.problems.example7 import Example7
-from lbc.problems.example8 import Example8
-from lbc.problems.example9 import Example9
 from lbc.util import sample_vector
 
 
-def get_problem_names():
-    return Problem.PROBLEM_MAP.keys()
-
-
-def get_problem(problem_name):
-    problem = Problem.PROBLEM_MAP[problem_name.lower()]
-    return problem
-
-
 class Problem:
-    PROBLEM_MAP = {
-        'example1': Example1,  # 2d single integrator regulator
-        'example2': Example2,  # 2d double integrator regulator
-        'example3': Example3,  # 3d dubins uncooperative tracking
-        'example4': Example4,  # 3d double integrator uncooperative tracking
-        'example5': Example5,  # game of atrition
-        'example6': Example6,  # bugtrap: 2d single integrator with obstacles
-        'example7': Example7,  # 2d double integrator uncooperative tracking
-        'example8': Example8,  # 2d single integrator pursuit evasion
-        'example9': Example9,  # homicidal chauffeur problem
-        'example10': Example10,  # dummy game problem
-        'example11': Example11,  # multiscale bugtrap: 2d single integrator with obstacles
-        'example12': Example12,  # modified homicidal chauffer
-        'example0_lbc': Example0_lbc
-    }
 
     def __init__(self):
         self.num_robots = None
