--- lbc/tests/__init__.py
@@ -1,16 +1,0 @@
-"""
-@title
-@description
-"""
-import argparse
-
-
-def main(main_args):
-    return
-
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='')
-
-    args = parser.parse_args()
-    main(vars(args))
